Remove extra multi-dropdown files
