--- app/javascript/app/components/dropdown/multi-dropdown/multi-dropdown-styles.scss
@@ -1,254 +1,0 @@
-@import '~styles/layout.scss';
-@import '~styles/utils.scss';
-
-$dd-height: 45px;
-
-.dropdown {
-  width: auto;
-  text-align: left;
-  margin-top: 3px;
-}
-
-.label {
-  display: block;
-  font-size: $font-size-s;
-  font-weight: $font-weight-bold;
-  color: $theme-color;
-  margin-bottom: 4px;
-}
-
-.infoButton {
-  margin-left: rem(7px);
-  height: 15px;
-  width: 15px;
-  min-width: 15px;
-  border-radius: 50%;
-
-  .infoIcon {
-    height: 8px;
-    width: 7px;
-  }
-}
-
-.isOpen {
-  .arrow {
-    transform: rotate(180deg);
-  }
-}
-
-.activeMark {
-  display: inline-block;
-  width: 10px;
-  height: 10px;
-  flex-shrink: 0;
-  border-radius: 50%;
-  background-color: $theme-color;
-}
-
-.childrenActiveMark {
-  display: inline-block;
-  width: 10px;
-  height: 10px;
-  flex-shrink: 0;
-  border-radius: 50%;
-  border: solid 2px $theme-color;
-}
-
-.disabled {
-  background-color: $light-gray;
-  cursor: not-allowed;
-}
-
-.container {
-  position: relative;
-}
-
-.placeholder {
-  color: $dark-gray;
-}
-
-.selector {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  position: relative;
-  border: solid 1px $border-color;
-  border-radius: 5px;
-  box-shadow: 0 1px 2px 0 rgba(0, 0, 0, 0.09);
-  height: $dd-height;
-  padding: 0 rem(10px);
-  cursor: pointer;
-  color: $theme-color;
-  z-index: 0;
-
-  &.alignLeft {
-    justify-content: flex-start;
-
-    .value {
-      padding-left: rem(18px);
-      padding-right: 0;
-    }
-  }
-
-  .value {
-    position: absolute;
-    top: calc(50% - 10px);
-    bottom: 0;
-    left: rem(10px);
-    right: rem(10px);
-    padding-right: rem(18px);
-    z-index: 0;
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    display: inline-block;
-
-    &.clearable {
-      padding-right: rem(28px);
-    }
-  }
-
-  input {
-    z-index: 1;
-    background-color: transparent;
-    width: 100%;
-    height: 100%;
-    border: 0;
-
-    &:focus {
-      outline: none;
-    }
-
-    &::placeholder {
-      color: $athens-gray;
-    }
-
-    &:read-only {
-      cursor: pointer;
-    }
-  }
-}
-
-.arrowBtn {
-  z-index: 3;
-  height: 100%;
-  cursor: pointer;
-
-  .arrow {
-    fill: $theme-color;
-    stroke: $theme-color;
-    width: 13px;
-    height: 13px;
-    stroke-width: 2px;
-    margin-right: rem(6px);
-  }
-}
-
-.clearBtn {
-  z-index: 3;
-  cursor: pointer;
-
-  .clearIcon {
-    max-width: rem(8px);
-    max-height: rem(8px);
-    margin-right: rem(16px);
-    fill: $dark-gray;
-  }
-}
-
-.menu {
-  border: solid 1px $border-color;
-  margin-top: -1px;
-  z-index: 1;
-  max-height: rem(200px);
-  overflow-y: scroll;
-  cursor: pointer;
-  color: $dark-gray;
-  position: absolute;
-  top: $dd-height;
-  width: 100%;
-  background-color: $white;
-
-  .item {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    width: 100%;
-    padding: rem(12px) rem(10px);
-    border-bottom: 1px solid $border-color;
-  }
-
-  .infoButton {
-    margin-right: rem(10px);
-    width: rem(16px);
-    max-width: rem(16px);
-    min-width: rem(16px);
-    height: rem(16px);
-    max-height: rem(16px);
-  }
-
-  .groupParent {
-    .item {
-      padding-right: rem(25px);
-      z-index: 5;
-    }
-
-    &.selected {
-      .item {
-        padding-right: rem(10px);
-        padding-left: rem(30px);
-      }
-    }
-
-    .groupIcon {
-      z-index: 6;
-    }
-  }
-
-  .groupIcon {
-    transform: rotate(-90deg);
-    position: absolute;
-    right: rem(10px);
-    fill: $dark-gray;
-
-    &:hover {
-      fill: $theme-color;
-    }
-
-    &.selected {
-      transform: rotate(90deg);
-      left: rem(10px);
-    }
-  }
-
-  .highlight {
-    &:hover {
-      color: $theme-color;
-    }
-  }
-}
-
-.itemWrapper {
-  justify-content: space-between;
-  align-items: center;
-  display: flex;
-  height: 0;
-  flex: 0;
-  margin-left: 100%;
-  width: 100%;
-  transition: none;
-  overflow: hidden;
-
-  &.base {
-    margin-left: -100%;
-    transition: margin-right 0.2s ease-in-out;
-  }
-
-  &.show {
-    height: auto;
-    margin-left: 0;
-    flex: 1;
-    overflow: initial;
-    transition: margin-left 0.2s ease-in-out;
-  }
-}
