+++ app/javascript/app/components/multi-dropdown/multi-dropdown-styles.scss
@@ -120,7 +120,7 @@ $dd-height: 45px;
     }
 
     &::placeholder {
-      color: $placeholder-gray;
+      color: $athens-gray;
     }
 
     &:read-only {
